# Optimize Rack::MediaType params and type parsers