```ruby
media_type: application/text

Warming up --------------------------------------
            original    38.980k i/100ms
           optimized    71.472k i/100ms
Calculating -------------------------------------
            original    482.257k (± 7.9%) i/s -      2.417M in   5.051485s
           optimized      1.010M (± 1.5%) i/s -      5.075M in   5.027185s

Comparison:
           optimized:  1009657.3 i/s
            original:   482256.8 i/s - 2.09x  slower

Calculating -------------------------------------
            original   280.000  memsize (     0.000  retained)
                         7.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)
           optimized   160.000  memsize (     0.000  retained)
                         4.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)

Comparison:
           optimized:        160 allocated
            original:        280 allocated - 1.75x more

media_type: application/text;CHARSET="utf-8"

Warming up --------------------------------------
            original    15.787k i/100ms
           optimized    20.540k i/100ms
Calculating -------------------------------------
            original    170.264k (± 1.1%) i/s -    852.498k in   5.007564s
           optimized    221.719k (± 4.5%) i/s -      1.109M in   5.015915s

Comparison:
           optimized:   221719.2 i/s
            original:   170264.4 i/s - 1.30x  slower

Calculating -------------------------------------
            original     1.112k memsize (     0.000  retained)
                        18.000  objects (     0.000  retained)
                         7.000  strings (     0.000  retained)
           optimized   912.000  memsize (     0.000  retained)
                        13.000  objects (     0.000  retained)
                         6.000  strings (     0.000  retained)

Comparison:
           optimized:        912 allocated
            original:       1112 allocated - 1.22x more
```